Updated Basic-Car-Maintenance-UITests
