+++ Basic-Car-Maintenance-UITests/BasicCarMaintenanceUITests.swift
@@ -9,18 +9,34 @@ import XCTest
 
 final class BasicCarMaintenanceUITests: XCTestCase {
 
-    override func setUpWithError() throws {
+    let app = XCUIApplication()
+    
+    override func setUp() {
         continueAfterFailure = false
-    }
-
-    override func tearDownWithError() throws {
-    }
-
-    func testExample() throws {
-        // UI tests must launch the application that they test.
-        let app = XCUIApplication()
         app.launch()
-
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
     }
+    
+    func testAddVehicle() {
+        // Navigate to the AddVehicleView
+        app.buttons["Settings"].tap()
+        app.buttons["Add Vehicle"].tap()
+        
+        // Fill in valid vehicle information
+        let nameTextField = app.textFields["Vehicle Name"]
+        nameTextField.tap()
+        nameTextField.typeText("My Car")
+        
+        let makeTextField = app.textFields["Vehicle Make"]
+        makeTextField.tap()
+        makeTextField.typeText("Toyota")
+        
+        let modelTextField = app.textFields["Vehicle Model"]
+        modelTextField.tap()
+        modelTextField.typeText("Camry")
+        
+        // Tap the "Add" button
+        app.buttons["Add"].tap()
+        
+    }
+    
 }
